+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - Operations on a `Root` have been renamed to have a `pathrs_inroot_` prefix.
   - `pathrs_open_root` has been renamed to `pathrs_open_root`, to avoid
     confusion with `pathrs_inroot_*` functions and clarify what it is opening.
+- python bindings: `Root.open` has been changed to be a wrapper of
+  `pathrs_inroot_open` instead of being a wrapper around the `Root`
+  constructor.
 
 ### Added ###
 - python bindings: add `Root.creat_raw` to create a new file and wrap it in a